+++ packages/block-editor/src/components/block-switcher/index.js
@@ -334,6 +334,7 @@ export const BlockSwitcher = ( { clientIds } ) => {
 						}
 						text={ blockIndicatorText }
 						toggleProps={ {
+							size: 'compact',
 							description: blockSwitcherDescription,
 							...toggleProps,
 						} }