+++ packages/components/src/toolbar/toolbar-group/toolbar-group-collapsed.tsx
@@ -28,6 +28,7 @@ function ToolbarGroupCollapsed( {
 			controls={ controls }
 			toggleProps={ {
 				...internalToggleProps,
+				size: 'compact',
 				'data-toolbar-item': true,
 			} }
 			{ ...props }